Improved decision engine and bug fixez
datatype bug, decision engine finds new periods, new API request URL

+++ src/main/java/ee/taltech/inbankbackend/endpoint/DecisionRequest.java
@@ -10,6 +10,6 @@ import lombok.Getter;
 @AllArgsConstructor
 public class DecisionRequest {
     private String personalCode;
-    private Long loanAmount;
+    private int loanAmount;
     private int loanPeriod;
 }
